Tweak the stellargraph Dockerfile for better caching (#590)
This uses the technique of installing dependencies in one layer, and then
copying the source code in another, so that changes to the source don't require
reinstalling all the dependencies. To do this effectively, it has to turn the
multi-stage build into a single stage one. However, this only increases the
compressed image size by 7MB (326MB -> 333MB) because the only thing that was
being excluded from the final image was the actual `stellargraph` source.

This then results in the majority of the image being shared, because the layer
containing the dependencies is rarely recomputed.

| situation        | branch   | build (s) | push (s) | total (s)    | compressed (MB) | uncompressed (MB) |
|------------------|----------|-----------|----------|--------------|-----------------|-------------------|
| same deps & code | this PR  | 5         | 7        | 60 (31+29)   | 0               | 0                 |
|                  | develop* | 150       | 73       | 341 (260+71) | 270             | 953               |
| same deps        | this PR  | 14        | 18       | 83 (52+31)   | 7.2             | 54.4              |
|                  | develop  | 150       | 73       | 341 (260+71) | 270             | 953               |
| different deps   | this PR  | 121       | 69       | 285 (210+75) | 277             | 1019              |
|                  | develop  | 150       | 73       | 341 (260+71) | 270             | 953               |

(* the computed installed dependencies are slightly different, so caching seems
to be ineffective, and thus this ends up being recomputed.)

The "total" time is the time to run the build step and then the publish
step. This captures some of the speed-up of sharing more layers, because a
significant amount of this time is the time required to download the images:
when the biggest layers are shared, most agents will have those layers available
and so not have to redownload them.

This includes a substitution for [the `no-cache`
setting](https://github.com/buildkite-plugins/docker-compose-buildkite-plugin#no-cache-optional-build-only),
so that we can have a scheduled build that refreshes the cache for any
dependency updates.