Convert Code Climate configuration to YAML (#596)
This converts our current Code Climate configuration to YAML without making any
changes. This was done by checking the GUI settings:
https://analysttools.slack.com/archives/C9YQZ8YVC/p1578265787024400

By versioning this with the code, we have all the benefits of git (tracking
changes, easy to revert, ...) as well as making easier to be tweaked as
required, rather than having to find someone with access to the web interface.