Add pip freeze on CI to ease debugging problems from version changes (#613)
This invocation shows the versions of each library installed with `pip`.

Doing this on CI makes it easy to find which version(s) of dependencies changed
between passing and failing builds, helping to narrow down the possible cause of
a failure. This thus ties into #611 and the daily scheduled build.

This switches to a non-editable install of `stellargraph` (`.`) because (a) CI
won't be doing any editing and so it's unnecessary, and (b) `pip freeze` tries
to invoke `git` for editable installs, which the CI docker images don't have.