Install tensorflow-cpu package when building docs on Read the Docs (#631)
With tensorflow 2.1, the `tensorflow` package includes GPU support by
default. This seems to be enough to exceed readthedocs's memory limits and cause
the build to fail. `tensorflow-cpu` 2.1 does not hit this problem, and so we can
fall back to that when building documentation.

We probably want to make this more generally configurable (#546), but that's a
more intricate problem, and this gets us back to having working documentation.

This also switches to only using the dependencies in `setup.py`, as building
documentation doesn't need any of the dev dependencies listed in
`requirements.txt`.

See: #602