Remove the metaclass from StellarGraph (#645)
This replaces the use of the metaclass with storing a `NetworkXStellarGraph` and
then calling its methods. This does not move the implementations directly into
the class, since that will be a lot of noise just to support the to-be-removed
`NetworkXStellarGraph` class better. Improvements to the functionality can land
directly in this class.

This copies the more experimental methods from `NetworkXStellarGraph` onto
`StellarGraph` too, so that no other changes are required to the code (other
than a couple of places that poke into the NetworkX internals (#599)).

See: #644