Refer to feature/bug templates in default zenhub issue template (#683)
Issues created via Zenhub defaults to using `.github/issue_template.md`, unlike
GitHub which just ignores it in favour of the templates in the
`.github/ISSUE_TEMPLATE/` folder, which are currently one for bugs and one
features. Those templates are more specific, do more work (add labels
automatically) and require less deletion of irrelevant information, and so it's
good for us to try to use them.

We can remind ourselves of this by replacing the Zenhub default template with a
reference to the dropdown that allow choosing the better templates.