Add @experimental to mark functions and classes as experimental (#685)
This adds a lightweight annotation that can be placed on a function, method or
class to automatically adjusts the documentation to include a big warning at the
top. The annotation requires a `reason=...` argument to be more specific about
why something is experimental, and (optionally) takes a list of issue numbers to
link in the `issues=...` argument.

For example, the docs for `StellarGraph.adjacency_types` start with:

> (!) Warning
>
> `StellarGraph.adjacency_types` is experimental: special-case method that needs
> more consideration. It may be difficult to use and may have major changes at
> any time.


And `RGCN` starts with:

> (!) Warning
>
> `RGCN` is experimental: it has severe known bugs (see: #649, #677). It may be
> difficult to use and may have major changes at any time.

This change also spells out a basic checklist for promoting something out of
"experimental".

Finally, to serve as examples, the annotation is used on some `StellarGraph`
methods, and also on the RGCN layers and generators (due to #649 and #677).

Why use an annotation? There's two basic approaches that we discussed on #617:

1. annotation (+ documentation)
2. `experimental` submodules (e.g. `stellargraph.layers.experimental`)

Despite some consensus on approach 2, this PR takes approach 1, because it can
be applied to individual methods, require fewer changes and file movements than
2 (as code moves in and out of experimental-ness), and, even if we take approach
2, having this annotation to automatically fill out the docs would still be
nice. =

However, approach 2 is more visible in code, as a reminder that things may
break: for example, `from stellargraph.layers.experimental import RGCN`. The
approach here relies on users noticing the annotation in the documentation.

Follow-up work may address this by adding dynamic warnings to experimental code.

See: #617