Remove some now unnecessary create_graph_schema calls (#701)
The list of node types is now directly available on a graph and so we do not
need to construct a schema to query it. For our current networkx graphs, this is
a full scan over all the nodes, which is suboptimal... but computing the graph
schema was strictly worse (multiple full scans over all nodes and all edges).