Remove GraphSchema node-type look-ups (#702)
There's basically the same amount of indirection for `StellarGraph.node_type` as
there was for `GraphSchema.get_node_type`, so this is not much slower. It makes
construction of a type like `GraphSAGELinkGenerator` ~20% slower than `develop`,
likely because the computation of the `edge_type_map` has to do two `dict`
look-ups rather than `dict`+`list`; as such, this slowdown is completely
recovered (and more) in #703 (the setup benchmarks end up ~30% faster than
`develop`). It seems to otherwise have no effect even on other benchmarks of
functionality that uses node-types, like `SampledHeterogeneousBreadthFirstWalk`.

In addition, the "GraphSchema" seems like it should be a summary of the graph,
rather than containing information about individual elements, and thus having
this per-node API in it is suboptimal.

This applies a "deprecation"/removal notice to future look-ups of these
attributes.

This simplification helps with #662, by reducing the API surface that has to be
reimplemented.