Remove NodeAttributeSpecification, replaced by pandas/sklearn (#707)
The `converter.py` file was reproducing a lot of functionality that's available
in pandas and sklearn, where it is likely more general, more efficient and more
reliable (since it's had more use). We've switched to doing using those
libraries in all our examples, and others should too.

Specifically:

- `BinaryConverter`: depending on the situation, potentially `some_data ==
  "some_value"`

- `CategorigalConverter `: `sklearn.preprocessing.OrdinalEncoder`
  <https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OrdinalEncoder.html>

- `OneHotCategorigalConverter`: `sklearn.preprocessing.OneHotEncoder`
  <https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OneHotEncoder.html>

- `NumericConverter`: `sklearn.preprocessing.scale`
  <https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.scale.html>

This PR is a "hard deprecation", since it removes all the functionality but
leaves the `NodeAttributeSpecification` class, emitting an error on
construction. I think we can remove it entirely after a single release.

There's some support functions that have been removed too.

See: #706