Remove GraphSchema edge-type look-ups (#703)
Follow on to #702 that does the same to edges as that did for nodes: removes the
per-element caching of edge types in the graph schema, since the graph schema
seems like it should represent a whole-graph summary rather than anything about
individual attributes.

This makes constructing a graph schema 60% faster because it has to do fewer
traversals of the graph. This thus recovers the performance that was lost in
#702 and more: this results in an overall speedup of 30% compared to `develop`
before #702 landed.

This applies a "deprecation"/removal notice to future look-ups of these
attributes, as well as to the `create_type_maps` parameter to
`create_graph_schema`.

This simplification helps with #662, by reducing the API surface that has to be
reimplemented.