Add numpy-based element backing stores for StellarGraph (#668)
This adds an implementation of the `StellarGraph` type using pandas and numpy
functionality.

See: #662

## Background

The implementation is driven by "ilocs" meaning "integer locations" or "internal
locations", similar to the pandas `.iloc` accessor. The implementation tries to
traffic in these as much as possible, to be able to store numpy arrays with
maximal efficient and minimal size. For example, sampling edges returns a list
of ilocs for the edges, which can be efficiently used to index other numpy
arrays to get any of the various edge properties (source, target, type, weight).

This represents a shift to a more columnar representation of graphs, which isn't
yet exposed in the API, but can be progressively propagated through the rest of
the library. Similarly, we can consider using ilocs instead of the external IDs
through the library too (e.g. edge sampling can typically sample directly from
the the compressed list of edge ilocs, without having to "dereference" them to
get the source or destination IDs).

This builds on the research work done by @geoffj-d61 and @adocherty in #520 (and
elsewhere).

## Status of the feature

This specific PR is not a complete replacement of the current networkx-based
`StellarGraph`:

- it's driven by separate arguments to the constructor (`nodes` and `edges`),
  and there's no automatic conversion from the old networkx graph (#715)
- there's limited testing, as only the benchmarks and a single example graph
  have been converted to use the new constructor (#717)
- there's some methods left unimplemented (ones that aren't called by the tests
  above) (#716)
- many of the methods are implemented inefficiently, to be close to the original
  networkx implementation: there's lots of room for optimisation, but doing it
  as future work will make it easier, when it requires API changes (#718)

The issues above track the follow up work to flesh these out, while keeping this
PR more focused.

Given the above, this lands as experimental, with `StellarGraph` still
completely supporting the old NetworkX-based implementation and using it by
default.

## Benchmarks

The only benchmarks that use the new class are the ones in
`test_stellargraph.py`. This section compares this PR with the networkx one on
develop.

### Performance

(Times in microseconds)

Getting neighbours is ~2× faster:

| code    | min |  max | mean | stddev | median | iqr |
|---------|----:|-----:|-----:|-------:|-------:|----:|
| this PR | 405 | 1210 |  476 |     88 |    435 |  73 |
| develop | 774 | 1990 |  918 |    149 |    882 | 124 |

Getting features for a set of node IDs is ~3× slower (I think this can be
partially addressed by switching to more iloc-based code):

| mode         | code    | min | max | mean | stddev | median | iqr |
|--------------|---------|----:|----:|-----:|-------:|-------:|----:|
| specify type | this PR | 100 | 446 |  117 |     30 |    107 | 8.0 |
|              | develop |  27 | 276 |   32 |     12 |     29 | 1.1 |
| infer type   | this PR | 179 | 670 |  207 |     48 |    187 |  24 |
|              | develop |  70 | 436 |   80 |     20 |     74 | 3.1 |

Creating a `StellarGraph` object is slower for small graphs, but seems to be
faster for larger ones:

0 nodes, 0 edges:

| features? | code    |  min |  max | mean | stddev | median |  iqr |
|-----------|---------|-----:|-----:|-----:|-------:|-------:|-----:|
| no        | this PR | 4900 | 8640 | 5490 |    558 |   5270 |  669 |
|           | develop |   19 |  277 |   24 |     11 |     21 | 0.93 |

100 nodes, 200 edges:

| features?     | code    |  min |   max | mean | stddev | median | iqr |
|---------------|---------|-----:|------:|-----:|-------:|-------:|----:|
| no            | this PR | 5320 |  9440 | 6220 |    778 |   6020 | 804 |
|               | develop | 1090 |  2710 | 1250 |    137 |   1220 | 120 |
| yes (dim=100) | this PR | 7840 | 11500 | 8760 |    670 |   8590 | 908 |
|               | develop | 1440 |  3120 | 1610 |    210 |   1530 | 197 |

1000 nodes, 5000 edges:

| features?     | code    |   min |    max |  mean | stddev | median |  iqr |
|---------------|---------|------:|-------:|------:|-------:|-------:|-----:|
| no            | this PR | 14000 |  20100 | 15700 |   1280 |  15500 | 2020 |
|               | develop | 27200 | 100000 | 36900 |  18900 |  30600 | 5800 |
| yes (dim=100) | this PR | 16800 |  22500 | 18800 |   1440 |  18500 | 1870 |
|               | develop | 30300 | 124000 | 42000 |  20500 |  37800 | 6610 |


### Memory

This uses significantly less memory for any graph of a reasonable size (in units
of k = 1000 bytes): for our 1000 node/5000 edge graph, the graph itself is
represented using 4× less memory, with almost exactly 1000 nodes * 100
features/node * 8 bytes/feature = 800k extra memory for the features themselves
(allowing this to not be copied is also future work).

0 nodes, 0 edges:

| features | code    | median |
|----------|---------|-------:|
| no       | this PR |  11.1k |
|          | develop |   2.3k |

100 nodes, 200 edges:

| features      | code    | median |
|---------------|---------|-------:|
| no            | this PR |  66.6k |
|               | develop |   143k |
| yes (dim=100) | this PR |   144k |
|               | develop |   231k |

1000 nodes, 5000 edges

| features      | code    | median |
|---------------|---------|-------:|
| no            | this PR |   747k |
|               | develop |  3169k |
| yes (dim=100) | this PR |  1539k |
|               | develop |  4009k |