Use numpy to accelerate create_graph_schema() (#743)
`create_graph_schema()` is a core method that powers a lot of other code, so
it's nice if it goes as fast as possible. We can do the same numpy analytics
that we did in `info` in #710. This now shares a lot of code with `info`.

It would probably be best to just have the `GraphSchema` store the counts
directly (rather than have `info` have to recompute them), however, doing this
now would require updating both the new StellarGraph and the NetworkX one, so
this improvement can wait for #715 and the removal of the latter.

The included benchmark uses on 1k nodes, 5k edges and says (times in
milliseconds):

| # types | code    |  min |   max | mean | stddev | median |  iqr |
|--------:|---------|-----:|------:|-----:|-------:|-------:|-----:|
|       1 | this PR |  967 |  2410 | 1160 |    140 |   1150 |  113 |
|         | develop | 7840 | 11300 | 8920 |    640 |   8980 | 1000 |
|       4 | this PR | 3450 |  4620 | 4080 |    179 |   4100 |  143 |
|         | develop | 7730 |  9900 | 8740 |    544 |   8680 |  650 |

That is, it's ~8× faster for homogeneous graphs, and 2× for a graph with 4 node
types (and thus likely `4 * 1 * 4 = 16` edge type triples, since the edges are
completely random).