Fix new StellarGraph _edge_weights method (#769)
This was completely broken: invalid code, since the `return weights[...]` line
was referencing an undefined `weights` variable. And, it was functionally
incorrect for undirected graphs, even if that was fixed to
`self._edges.weights`.

This rewrites the code to find the set of edges between two nodes by
intersecting their adjacency lists. This is somewhat suboptimal, but good enough
to unblock the current implementation, and, the users of this method should
probably be rewritten to use a different approach (a more vectorised one: #718)
anyway. In many cases, it's likely to be querying information about edges that
were selected by some earlier process: that selection process can/should just
include the weight information directly.

See: #768