Use an integer-typed array for isolated nodes in new StellarGraph (#771)
The old form was from before the new `StellarGraph` element-data storage
switched to using "ilocs" everywhere, and so was set to have the same type as
the IDs themselves. However, now that the code has switched to ilocs, the
`edge_ilocs` method should always return an array with an iloc-esque dtype, that
is, an array that can be used to index other arrays.

The choice of `np.uint8` as the dtype is to be as compatible as possible,
require minimal type promotions when two arrays of integer dtypes are used
together, e.g. the following example does an operation on an "empty_iloc" array,
with other arrays of different types, and we can see the resulting dtypes:

```python
>>> empty_ilocs = np.array([], dtype=np.uint8)
>>> np.intersect1d(empty_ilocs, np.array([], dtype=np.uint16))
array([], dtype=uint16)
>>> np.intersect1d(empty_ilocs, np.array([], dtype=np.int64))
array([], dtype=int64)
```

See: #770