Only accept one node in new StellarGraph.node_type (#777)
Previously this was trying to allow batch queries, but this differed to the old
implementation. The new implementation would always return a sequence
(pandas.Index) of node types, even if only a single node was queried. This PR
switches the new StellarGraph `node_type` method to behave like the old one:
accept a single node ID and return a single node type. We can move to batched
queries in future (this relates to #718).

See: #775