Have explicitly-cached docker images for running on CI (#788)
This uses the Buildkite docker-compose plugin to eliminate the 1-2 minutes of
"install dependencies" required to run tests and demos. It does this by
prebuilding a docker image that contains the dependencies required for this, in
a way that benefits from caching, and so only has to be recomputed if the
dependencies themselves change (approximated by whether `setup.py` changes). In
addition, by using docker's caching infrastructure, we automatically have cache
invalidation managed for us: when dependencies are added, removed or modified in
`setup.py`, the image will recreated.

The image is called `stellargraph-ci-runner`
https://hub.docker.com/repository/docker/stellargraph/stellargraph-ci-runner and
is public. It's quite large (605 MB compressed), but changes rarely, and so is
likely to end up already cached on our build agents, and thus the time to
pull/push is minimal (just the time to check that the layers already exist).

Because we test against both 3.6 and 3.7, this pushes an image for each of them
under `3_6-<normal tag>` and `3_7-<normal tag>` (examples: `3_6-PR-788`,
`3_7-latest`). Because this is just a cache image, the image is pushed under
that final tag as soon as it is built (that is, the prebuild step includes
`push: ...`), rather than done as part of the final publish after all testing.

This makes the individual steps noticeably faster (about 90s), e.g.

| step                                             | develop (m:ss) | this PR (m:ss) |
|--------------------------------------------------|---------------:|---------------:|
| test 3.6                                         |           8:19 |           6:42 |
| test 3.7                                         |           8:00 |           6:35 |
| `calibration-pubmed-link-prediction`             |           4:10 |           2:58 |
| `attri2vec-citeseer-node-classification-example` |           5:02 |           3:32 |
| `sgc-node-classification-example`                |           2:52 |           1:26 |

develop build: https://buildkite.com/stellar/stellargraph-public/builds/1439
this PR build: https://buildkite.com/stellar/stellargraph-public/builds/1447

This doesn't make the end-to-end build time that much faster with a build with
no waiting: the prebuild step takes about 40s-65s (~20s of setup, 20s of pulling
the original image if the agent doesn't have it already, 5s of building the
image from cache, 20s of pushing the unchanged image twice), and only then can
the downstream steps start, and that's similar to the ~90s required to install
dependencies before testing starts.

On the other hand, it does dramatically reduce the agent time required, per
build:

- t2medium: 2×30 seconds = 1 minute, and
- t2large: 24 (number of notebooks that are actually tested)×90 seconds = 36 minutes

This likely also means that there's less waiting-for-agents (because agents are
spending less time working, and able to continue onto the next job sooner) and
so builds are faster in practice.

There's still some future work building on this, that will be done as separate
PRs since they're not part of this core change:

- install `papermill` into the CI image (cuts another ~15s off the installation
  time for notebooks, i.e. another 6 minutes of t2large agent time saved) (#878)
- have the notebook formatting check run as a separate parallel step, outside
  the 3.6 testing one (#877)