Move `load_dataset_BlogCatalog3` to a `BlogCatalog3.load` method (#888)
This moves the existing `load_dataset_BlogCatalog3` function to be part of the
new `datasets` module, and tweaks it to return a `StellarGraph` object instead
of a NetworkX graph. The old function now emits deprecation warning and then
calls through to the new code (and does a conversion to a NetworkX graph).

This is a start on #812.