Fix GAT interpretability demo for new StellarGraph order changes (#880)
This does a small change to `node-link-importance-demo-gat.ipynb` to get it
working again after the switch to the new `StellarGraph` class. In particular,
the change to the new StellarGraph changed the order in which nodes appear. so:

1. the code that was (effectively) `G.nodes()[7]` was no longer selecting the
   node with ID `1109199`
2. the code that was translating between IDs and indices was sometimes doing it
   via the `Gnx` (the NetworkX graph) and sometimes via `G`/`graph_nodes` (the
   `StellarGraph`) which gave different results.

This switches 1 to starting with the actual node ID of interest, and 2 to always
using a single list (`graph_nodes` or `nodes` as appropriate).

The code is quite poor in many ways here (all of the `.index` calls are O(n)),
but in the interests of getting this fixed as soon as possible, this tries to
keep the changes here minimal and not make any improvements. The work to get
this notebook running on CI is a little intricate/fiddly (#881).

See: #868