Implement the ComplEx knowledge graph algorithm (#848)
This implements the ComplEx knowledge graph algorithm from: Complex Embeddings
for Simple Link Prediction; Théo Trouillon, Johannes Welbl, Sebastian Riedel,
Éric Gaussier and Guillaume Bouchard, ICML 2016. 
http://jmlr.org/proceedings/papers/v48/trouillon16.pdf

This comes in three parts:

1. a `KGTripleGenerator` where the `flow` method takes a list of `(source, edge
   type, target)` triples (aka `(subject, relation, object)`) and yields them as
   batches of sequential integers to use to look-up in embedding layers
2. the `ComplEx` model which creates the appropriate embedding layers and
   applies the ComplEx scoring mechanism to them (which gives logit
   "probabilities" for a link)
3. some minor changes to `StellarGraph` and its supporting code, so that passing
   in an unknown node gets an appropriate error

This doesn't add a notebook that reproduces the paper, because the datasets used
in it are new to the library, and the MRR metrics used are awkward to
compute. This can/should be future work (#862). However it does add tests,
include one that does simple and manual computation of the scores for links, to
validate that it matches the paper's description.

More future work: There's a whole class of knowledge graph algorithms that have
a similar approach (e.g. DistMult #755), just changing the details of the
embeddings (basically real or complex, matrix or vector) and the scoring
mechanism. The `KGTripleGenerator` class can be easily reused for them, and
there's likely some commonalities that we can factor out of the `ComplEx` model
to share code. In the interests of YAGNI and waiting to choose the right
generalisation, this PR does not try to generalise the `ComplEx` model yet. The
ComplEx paper has a good summary of these similar algorithms in Table 1 on page
3; it includes algorithms like TransE, DistMult and RESCAL.

See: #756