Streamline and redocument the new StellarGraph constructor (#891)
This makes the final changes to the `StellarGraph` constructor that we need to
consider it non-experimental. This effectively deprecates the NetworkX form
(e.g. `StellarGraph(g_nx, node_features=...)`), replaced by
`StellarGraph.from_networkx`, and moves the new form front-and-centre.

Summary of changes:

- moves where the Pandas node and edge DataFrames are passed: they were late in
  the argument list and so required keywords `StellarGraph(nodes=n, edges=e)`,
  but are now first, and so passing as positional arguments is possible
  `StellarGraph(n, e)` (keywords still works, though).
- tweaks some argument names to make more sense with the new form
  (`edge_weight_label` -> `edge_weight_column`).
- forces everything other than the nodes and edges arguments to be passed using
  keyword arguments, e.g. `StellarGraph(n, e, True)` is an error, it has to be
  `StellarGraph(n, e, is_directed=True)`. This gives us more flexibility for
  changes in future.
- entirely rewrites the class documentation to reflect the new form.
- removes the experimental warning from the code (and from the docs).

The NetworkX form does not yet give a `DeprecationWarning`, because that would
cause a significant chunk of our tests and examples to emit these warnings; once
we've made progress on #717, we can add a proper warning here to encourage
people to migrate.

Is changing the argument order, and forcing keyword arguments safe (that is,
would it break any existing code)? I think it is safe. All of our tests and
examples use keyword arguments for everything except the graph (first argument),
so these do not break, and it seems like this has encouraged all of our users to
do the same: based on a moderate sized sample, no (open source) code does
`StellarGraph(g_nx, True, "weight_label", "node_type")` or similar. I did this
sample by:

- searching in our dependent code
  https://github.com/stellargraph/stellargraph/network/dependents?package_id=UGFja2FnZS0xNjI2NTY5NDU%3D
  State at the time of committing: `zdjohn/freshgraph`,
  `alphagov/govuk-network-embedding`, `amaiya/ktrain`,
  `ddxgz/linkedinfo-ml-models`.

- reviewing the first 10 or so pages of a GitHub search for `StellarGraph`
  https://github.com/search?q=StellarGraph+-is%3Afork+-org%3Astellargraph&type=Code

#893 does some similar work to clean-up `from_networkx`.

See: #723