Add buildkite annotation when notebook parallelism setting is wrong (#916)
If the number of notebooks doesn't match the `parallelism` setting used for
testing notebooks, all of those steps will fail, and one needs to go into them
to deduce what the problem was, and then interpret the message. This PR (a)
rewrites the message to be a bit clearer, including mentioning the common case
of needing to merge with develop, and (b) adds it as an annotation
(https://buildkite.com/docs/agent/v3/cli-annotate) at the top of the build for
convenience.