Add a 'plot_history' function for viewing metrics across epochs (#902)
This factors out the `plot_history` function that has been, essentially,
copy-and-pasted into every notebook into a function in the library. It places it
in `stellargraph.utils.plot_history` (in `utils` because there's not an obvious
better place at the moment: related to #909)

The plot allows for multiple histories (for ensemble models) and validation
history (optional). It puts all of the plots in a column, sharing the x-axis
(epoch number).

This doesn't rerun all of the notebooks that were updated:

- `rgcn-aifb-node-classification-example.ipynb` fails: #911
- `hateful-twitters.ipynb` fails: #912
- `directed-graphsage-on-cora-neo4j-example.ipynb` and
  `undirected-graphsage-on-cora-neo4j-example.ipynb`: I haven't got Neo4j
  running locally (relates to #849 and #904)

See: #898