Update release procedure
Major update to release procedure here is the change in the order
of steps - PyPi and Conda publishing should now be done before
pushing to master. This is so that any problems with the publishing
steps can be resolved before packaging the release on GitHub.

Some of the wording and formatting also updated in the release
procedure document for clarity.