Optimise `setdiag(0)` in StellarGraph.to_adjacency_matrix on undirected graphs (#932)
On a CSR matrix
(https://docs.scipy.org/doc/scipy/reference/generated/scipy.sparse.csr_matrix.html),
`setdiag` will change the sparsity structure, meaning it has to introduce new
elements into the optimised `numpy` arrays that it stores under the hood (which
is `O(edges)` in our case). This is true even when setting the diagonal to `0`
which theoretically only needs to clear the elements that exist, not add new
ones (https://github.com/scipy/scipy/issues/11600). Per that issue, it's faster
to manually find the zero elements in the diagonal, and clear only them to avoid
changing the sparsity structure, which is `O(nodes)`.

This improves the included benchmark on undirected graphs with 1000 nodes and
5000 edges significantly (15&times;), and will likely have a large improvement
on larger graphs. (Times in microseconds.)

| type       | code    | min (s) | mean (s) | stddev (s) |
|------------|---------|--------:|---------:|-----------:|
| directed   | this PR |     484 |      565 |         88 |
| directed   | develop |     485 |      573 |         92 |
| undirected | this PR |     937 |     1060 |        155 |
| undirected | develop |   12900 |    15100 |       1240 |

(The directed code path doesn't change and those observations are within noise
of each other.)

This also switches to `transpose` rather than constructing a whole new separate
matrix for the backwards links, because that's also faster (min time: 1110us ->
937us).

See: #931