Use str IDs for RGCN notebook, not rdflib types (#983)
Previously the notebook wouldn't run because numpy was getting confused by the
`rdflib` term types that were being used as the node IDs before; in particular,
`rdflib.term.URIRef` and `rdflib.term.Literal` are subclasses of `str` with a
custom hash and equality. Creating an array from a collection of such IDs gets
treated as an array of strings (for instance, RGCN's generator feeds
`list(graph.nodes())` into `StellarGraph.node_features` which does `np.asarray`
internally). However, the types get preserved as-is elsewhere (in particular, in
a pandas Series or Index) and so there's a mismatch between the `rdflib` types
and `str` when doing node look-ups into hash table later: the `str` values don't
behave like the `rdflib` ones.

Arguably, just converting these types to proper strings isn't the correct fix,
but this is an acceptable short-term fix to allow the notebook to work. #988
covers handling this problem more generally.

See: #911, #972