Add feature vector dimension info to StellarGraph.info (#979)
This provides more insight into the shape of a dataset. For example, `info` on
Cora now looks like:

```
StellarGraph: Undirected multigraph
 Nodes: 2708, Edges: 5429

 Node types:
  paper: [2708]
    Features: float32 vector, length 1433
    Edge types: paper-cites->paper

 Edge types:
    paper-cites->paper: [5429]
```

and on MovieLens:

```
StellarGraph: Undirected multigraph
 Nodes: 2625, Edges: 100000

 Node types:
  movie: [1682]
    Features: float32 vector, length 19
    Edge types: movie-rating->user
  user: [943]
    Features: float32 vector, length 24
    Edge types: user-rating->movie

 Edge types:
    user-rating->movie: [100000]
```

Note the `Features: ...` lines.

This does not run all 28 of the notebooks that call `StellarGraph.info()`, since
this is an optional piece of information and rerunning all of them will either
take a long time & result in a whole pile of other changes (for running the
whole thing), or will be fiddly (for rerunning on the cell that calls
`info`). It's ok for these notebooks to be updated progressively as they are
re-run for other larger changes.

See: #333