Add CiteSeer.load (#989)
This adds `CiteSeer.load` for loading that dataset into a `StellarGraph` object.

The dataset is essentially the same as Cora, in that it's from the same source
and has a very similar structure. It thus shares the code. The only
differences/new code are:

- a different number of features (set by `_NUM_FEATURES`)
- some node IDs are strings (set by `_NODES_DTYPES`)
- there are some dangling edges, where the edges refer to nodes that are not
  included in the list of feature vectors in the `citeseer.content` file (these
  are always ignored)

This PR also uses it in all two of notebooks that use the CiteSeer dataset:
`stellargraph-attri2vec-citeseer.ipynb`,
`attri2vec-citeseer-node-classification-example.ipynb`.

Similarly to #977, this uses a function to share functionality, rather than a
subclass, because it's probably not worth the additional complexity of having
another layer at the moment, as well as needing to adjust things so the
`__subclasses__` works correctly in the tests.

See: #812