Check info() against literal expected output (#994)
The `test_info_heterogeneous` test for `StellarGraph.info` on a heterogeneous
graph was incorrect, because did not have some assertions when it should've.

This fixes that by checking the exact output for `info()`, meaning the full
string including indentation/new lines. This is good because it means we're
validating what the users see, and so will catch any mistakes that break the
formatting.