Test neo4j notebooks on CI (#1046)
This uses docker-compose to add a Neo4j instance running on CI. The neo4j
container is created by building a docker image from the official neo4j images
that includes the "APOC" plugin (https://neo4j.com/developer/neo4j-apoc/) that
we need. Building and running this image is easy and fast, only a few seconds if
the neo4j image is already available on the agent; as such, for simplicity, this
unconditionally runs the image neo4j for every step that uses one of the 3.x
runners, even though most do not use neo4j.

The notebooks are tested as a single step, that runs them in series, because the
directed GraphSAGE notebooks rely on having Cora already available in the
database, and thus the `load-cora-into-neo4j.ipynb` notebook needs to run first.

This makes some minor edits to the notebooks so that they can work on CI, with
the major one restructuring how the connection to neo4j works, by using
Neo4j/py2neo's defaults for everything except the host, which is read from the
`STELLARGRAPH_NEO4J_HOST` environment variable if it exists. This is used on CI
to tell to the code to connect to the neo4j container which is available at
`neo4j`, not `localhost` since it's a separate container. An alternative would
be to specify `network_mode: "service:runner-3_6"` in `neo4j`, so that it's
available under the runner container's `localhost` and thus need no custom code
in the notebook at all. Thoughts?

This tests on Neo4j 3.5, which is not the latest release, but (a) it's probably
good to support a bit of a range, and (b) the notebooks fail on Neo4j 4.0
(#1055). Once that issue is fixed, I believe we should be able to extend this to
test against multiple versions of Neo4j using build args
(https://docs.docker.com/engine/reference/builder/#arg).

See: #849